Corrigido problema no campo rg ao editar pessoa física; Refs #2686

+++ ieducar/intranet/atendidos_cad.php
@@ -337,7 +337,7 @@ class indice extends clsCadastro
       'inline'      => true
     );
 
-    $this->inputsHelper()->numeric('rg', $options);
+    $this->inputsHelper()->integer('rg', $options);
 
 
     // data emissão rg
